requirement.txt
django
psycopg2
django-ratings
pinax-ratings
numpy
django-bootstrap3
django-filter
